fix: harden codec/transport and close Java multi-matcher parity gap
Pre-release robustness pass on the client's networking/encoding layer plus
one real parity fix, all surfaced by a manual+automation QA review.

- Path injection: build admin URLs via percentEncodedPath and route
  user-supplied scenario/file names through AdminClient.pathSegment, which
  percent-encodes `/?#` and rejects an empty segment (new
  WireMockError.invalidArgument) — a name can no longer inject path
  segments, traverse, or bleed into the query/fragment.
- register(raw:) now validates JSON up front then sends the user's exact
  bytes, instead of parse-and-re-encode through JSONValue (which reordered
  keys and coerced numbers) — a "raw" hatch must be byte-faithful.
- Non-UTF8 error bodies fall back to ISO-8859-1 instead of collapsing to an
  empty string, so a server error is never silently dropped.
- init?(scheme:host:port:) validates port 1...65535 and a non-blank host.
- Journal loggedDate and SubEvent timeOffsetNanos widened to Int64.
- Parity: repeated withHeader/withQueryParam/withCookie/withPathParam/
  withFormParam on one key now AND-combine (flattened) instead of the later
  call overwriting the earlier, matching Java WireMock. The wire shape stays
  one matcher per key ({"and":[...]}), which the 3.13.2 server accepts.

+++ Sources/WireMock/Models/Journal.swift
@@ -19,7 +19,8 @@ public struct LoggedRequest: Codable, Sendable, Hashable {
     public var cookies: [String: HeaderValue]?
     public var body: String?
     public var bodyAsBase64: String?
-    public var loggedDate: Int?
+    /// Epoch milliseconds (~1.7e12) — `Int64` so it can't overflow a 32-bit `Int`.
+    public var loggedDate: Int64?
     public var loggedDateString: String?
     public var queryParams: JSONValue?
     public var formParams: JSONValue?
@@ -58,7 +59,8 @@ public struct Timing: Codable, Sendable, Hashable {
 /// free-form JSON (the server puts a diff/near-miss report here).
 public struct SubEvent: Codable, Sendable, Hashable {
     public var type: String?
-    public var timeOffsetNanos: Int?
+    /// Nanosecond offset — `Int64` to stay safe from 32-bit overflow.
+    public var timeOffsetNanos: Int64?
     public var data: JSONValue?
 }
